+++ src/contactForm.js
@@ -1,8 +1,9 @@
 class RecrasContactForm {
     constructor(options = {}) {
+        this.datePicker = null;
         this.languageHelper = new RecrasLanguageHelper();
 
-        if ((options instanceof RecrasOptions) === false) {
+        if (!(options instanceof RecrasOptions)) {
             throw new Error(this.languageHelper.translate('ERR_OPTIONS_INVALID'));
         }
         this.options = options;
@@ -82,8 +83,9 @@ class RecrasContactForm {
         if (this.hasCountryField()) {
             waitFor.push(this.getCountryList());
         }
-        if (this.hasPackageField()) {
-            waitFor.push(this.getPackages(this.options.getFormId()));
+        if (this.hasDateField()) {
+            waitFor.push(RecrasCalendarHelper.loadScript());
+            RecrasCSSHelper.loadCSS('pikaday');
         }
         return Promise.all(waitFor).then(() => {
             let html = '<form class="recras-contactform">';
@@ -124,6 +126,7 @@ class RecrasContactForm {
         return this.fetchJson(this.options.getApiBase() + 'contactformulieren/' + this.options.getFormId() + '?embed=Velden')
             .then(form => {
                 this.contactFormFields = form.Velden;
+                this.packages = this.sortPackages(form.Arrangementen);
                 return this.contactFormFields;
             });
     }
@@ -136,35 +139,34 @@ class RecrasContactForm {
             });
     }
 
-    getPackages(contactFormID) {
-        return this.fetchJson(this.options.getApiBase() + 'contactformulieren/' + contactFormID)
-            .then(json => {
-                this.packages = json.Arrangementen.sort((a, b) => {
-                    // Prioritise package name
-                    if (a.arrangement < b.arrangement) {
-                        return -1;
-                    }
-                    if (a.arrangement > b.arrangement) {
-                        return 1;
-                    }
-                    // Sort by ID in the off chance that two packages are named the same
-                    if (a.id < b.id) {
-                        return -1;
-                    }
-                    if (a.id > b.id) {
-                        return 1;
-                    }
-                    // This cannot happen
-                    return 0;
-                });
-                return this.packages;
-            });
+    sortPackages(packs) {
+        return packs.sort((a, b) => {
+            // Prioritise package name
+            if (a.arrangement < b.arrangement) {
+                return -1;
+            }
+            if (a.arrangement > b.arrangement) {
+                return 1;
+            }
+            // Sort by ID in the off chance that two packages are named the same
+            if (a.id < b.id) {
+                return -1;
+            }
+            if (a.id > b.id) {
+                return 1;
+            }
+            // This cannot happen
+            return 0;
+        });
     }
 
     hasFieldOfType(identifier) {
         return this.contactFormFields.filter(field => {
             return field.field_identifier === identifier;
         }).length > 0;
+    }
+    hasDateField() {
+        return this.hasFieldOfType('boeking.datum');
     }
     hasCountryField() {
         return this.hasFieldOfType('contact.landcode');
@@ -201,10 +203,15 @@ class RecrasContactForm {
             return `<h3>${ field.naam }</h3>`;
         }
 
+        const today = RecrasDateHelper.toString(new Date());
+        const datePattern = '[0-9]{4}-(0[1-9]|1[012])-(0[1-9]|1[0-9]|2[0-9]|3[01])';
+        const timePattern = '(0[0-9]|1[0-9]|2[0-3])(:[0-5][0-9])';
+
         let label = this.showLabel(field, idx);
         let attrRequired = field.verplicht ? 'required' : '';
         let classes;
         let html;
+        let placeholder;
         let fixedAttributes = `id="contactformulier-${ idx }" name="contactformulier${ idx }" ${ attrRequired } data-identifier="${ field.field_identifier }"`;
         switch (field.soort_invoer) {
             case 'contactpersoon.geslacht':
@@ -230,7 +237,7 @@ class RecrasContactForm {
             case 'contact.soort_klant':
                 html = `<div class="radioGroup">`;
                 field.mogelijke_keuzes.forEach(choice => {
-                    html += `<label><input type="radio" name="contactformulier${ idx }" value="${ choice }"${ attrRequired } data-identifier="${ field.field_identifier }">${ choice }</label>`;
+                    html += `<label><input type="radio" name="contactformulier${ idx }" value="${ choice }" ${ attrRequired } data-identifier="${ field.field_identifier }">${ choice }</label>`;
                 });
                 html += `</div>`;
                 return label + html;
@@ -257,18 +264,19 @@ class RecrasContactForm {
             case 'contact.landcode':
                 html = `<select ${ fixedAttributes } autocomplete="country">`;
                 Object.keys(this.countries).forEach(code => {
-                    let selectedText = code.toUpperCase() === this.languageHelper.getCountry() ? ' selected' : '';
-                    html += `<option value="${ code }"${ selectedText }>${ this.countries[code] }`;
+                    let selectedText = code.toUpperCase() === this.languageHelper.getCountry() ? 'selected' : '';
+                    html += `<option value="${ code }" ${ selectedText }>${ this.countries[code] }`;
                 });
                 html += '</select>';
                 return label + html;
-            case 'boeking.datum': //TODO: add optional fallback
-                const today = RecrasDateHelper.toString(new Date());
-                return label + `<input type="date" ${ fixedAttributes } min="${ today }" pattern="[0-9]{4}-(0[1-9]|1[012])-(0[1-9]|1[0-9]|2[0-9]|3[01])">`;
+            case 'boeking.datum':
+                placeholder = this.languageHelper.translate('DATE_FORMAT');
+                return label + `<input type="text" ${ fixedAttributes } min="${ today }" placeholder="${ placeholder }" pattern="${ datePattern }" autocomplete="off">`;
             case 'boeking.groepsgrootte':
                 return label + `<input type="number" ${ fixedAttributes } min="1">`;
-            case 'boeking.starttijd': //TODO: add optional fallback
-                return label + `<input type="time" ${ fixedAttributes } pattern="(0[0-9]|1[0-9]|2[0-3])(:[0-5][0-9])">`;
+            case 'boeking.starttijd':
+                placeholder = this.languageHelper.translate('TIME_FORMAT');
+                return label + `<input type="time" ${ fixedAttributes } placeholder="${ placeholder }" pattern="${ timePattern }" step="300">`;
             case 'boeking.arrangement':
                 const preFilledPackage = this.options.getPackageId();
 
@@ -297,6 +305,21 @@ class RecrasContactForm {
             .then(html => {
                 this.appendHtml(html);
                 this.findElement('.recras-contactform').addEventListener('submit', this.submitForm.bind(this));
+                if (this.hasDateField()) {
+                    let pikadayOptions = Object.assign(
+                        RecrasCalendarHelper.defaultOptions(),
+                        {
+                            field: this.findElement('[data-identifier="boeking.datum"]'),
+                            i18n: RecrasCalendarHelper.i18n(this.languageHelper),
+                            numberOfMonths: 1,
+                            toString(date, _format) {
+                                return RecrasDateHelper.datePartOnly(date);
+                            },
+                        }
+                    );
+
+                    this.datePicker = new Pikaday(pikadayOptions);
+                }
                 this.loadingIndicatorHide();
             });
     }