+++ src/booking.js
@@ -1,6 +1,6 @@
 /*******************************
 *  Recras integration library  *
-*  v 1.0.0                     *
+*  v 1.1.0                     *
 *******************************/
 
 class RecrasBooking {