+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## branch 21
+## 1.1.0 (2019-09-25)
 * Optimise loading of contact form that contains a "package" field
 * Add datepicker to date field in contact forms